backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.86.0
python-dotenv==1.0.0
functions-framework>=3.0.0
stripe==7.6.0
pydantic-settings==2.1.0
supabase==2.10.0
